moved bridging code into its own file
