+++ holochain.go
@@ -9,7 +9,6 @@ package holochain
 import (
 	"bytes"
 	"encoding/gob"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"github.com/google/uuid"
@@ -673,183 +672,5 @@ func (h *Holochain) VerifySignature(signature []byte, data string, pubKey ic.Pub
 	if err != nil {
 		return
 	}
-	return
-}
-
-type BridgeSpec map[string]map[string]bool
-
-// AddBridgeAsCallee registers a token for allowing bridged calls from some other app
-// and calls bridgeGenesis in any zomes with bridge functions
-func (h *Holochain) AddBridgeAsCallee(fromDNA Hash, appData string) (token string, err error) {
-	Debugf("Adding bridge to %s from %v with appData: %s", h.nucleus.dna.Name, fromDNA, appData)
-	err = h.initBridgeDB()
-	if err != nil {
-		return
-	}
-	var capability *Capability
-
-	bridgeSpec := h.makeBridgeSpec()
-	var bridgeSpecB []byte
-
-	if bridgeSpec != nil {
-		bridgeSpecB, err = json.Marshal(bridgeSpec)
-		if err != nil {
-			return
-		}
-	}
-
-	capability, err = NewCapability(h.bridgeDB, string(bridgeSpecB), nil)
-	if err != nil {
-		return
-	}
-
-	for zomeName, _ := range bridgeSpec {
-		var r Ribosome
-		r, _, err = h.MakeRibosome(zomeName)
-		if err != nil {
-			return
-		}
-		Debugf("Running BridgeTo Genesis for %s", zomeName)
-		err = r.BridgeGenesis(BridgeTo, fromDNA, appData)
-		if err != nil {
-			return
-		}
-	}
-
-	token = capability.Token
-
-	return
-}
-
-func (h *Holochain) initBridgeDB() (err error) {
-	if h.bridgeDB == nil {
-		h.bridgeDB, err = buntdb.Open(filepath.Join(h.DBPath(), BridgeDBFileName))
-	}
-	return
-}
-
-func checkBridgeSpec(spec BridgeSpec, zomeType string, function string) bool {
-	f, ok := spec[zomeType]
-	if ok {
-		_, ok = f[function]
-	}
-	return ok
-}
-
-func (h *Holochain) makeBridgeSpec() (spec BridgeSpec) {
-	var funcs map[string]bool
-	for _, z := range h.nucleus.dna.Zomes {
-		for _, f := range z.BridgeFuncs {
-			if spec == nil {
-				spec = make(BridgeSpec)
-			}
-			_, ok := spec[z.Name]
-			if !ok {
-				funcs = make(map[string]bool)
-				spec[z.Name] = funcs
-
-			}
-			funcs[f] = true
-		}
-	}
-	return
-}
-
-// BridgeCall executes a function exposed through a bridge
-func (h *Holochain) BridgeCall(zomeType string, function string, arguments interface{}, token string) (result interface{}, err error) {
-	if h.bridgeDB == nil {
-		err = errors.New("no active bridge")
-		return
-	}
-	c := Capability{Token: token, db: h.bridgeDB}
-	var bridgeSpecStr string
-	bridgeSpecStr, err = c.Validate(nil)
-	if err == nil {
-		if bridgeSpecStr != "*" {
-			bridgeSpec := make(BridgeSpec)
-			err = json.Unmarshal([]byte(bridgeSpecStr), &bridgeSpec)
-			if err == nil {
-				if !checkBridgeSpec(bridgeSpec, zomeType, function) {
-					err = errors.New("function not bridged")
-					return
-				}
-			}
-		}
-		if err == nil {
-			result, err = h.Call(zomeType, function, arguments, ZOME_EXPOSURE)
-		}
-	}
-
-	if err != nil {
-		err = errors.New("bridging error: " + err.Error())
-
-	}
-
-	return
-}
-
-// AddBridgeAsCaller associates a token with an application DNA hash and url for accessing it
-// it also runs BridgeGenesis for the From side
-func (h *Holochain) AddBridgeAsCaller(toDNA Hash, token string, url string, appData string) (err error) {
-	Debugf("Adding bridge from %s to %v with appData: %s", h.nucleus.dna.Name, toDNA, appData)
-	err = h.initBridgeDB()
-	if err != nil {
-		return
-	}
-	toDNAStr := toDNA.String()
-	err = h.bridgeDB.Update(func(tx *buntdb.Tx) error {
-		_, _, err = tx.Set("app:"+toDNAStr, token, nil)
-		if err != nil {
-			return err
-		}
-		_, _, err = tx.Set("url:"+toDNAStr, url, nil)
-		if err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		return
-	}
-
-	// TODO  possible that we shouldn't add the bridge unless the there is some Zome with BridgeTo?
-	// the way this is is just that the only way to get the from genesis to run is if it's set
-	for _, z := range h.nucleus.dna.Zomes {
-		if z.BridgeTo.String() == toDNAStr {
-			var r Ribosome
-			r, _, err = h.MakeRibosome(z.Name)
-			if err != nil {
-				return
-			}
-			Debugf("Running BridgeFrom Genesis for %s", z.Name)
-			err = r.BridgeGenesis(BridgeFrom, toDNA, appData)
-			if err != nil {
-				return
-			}
-
-		}
-	}
-	return
-}
-
-var BridgeAppNotFoundErr = errors.New("bridge app not found")
-
-// GetBridgeToken returns a token given the a hash
-func (h *Holochain) GetBridgeToken(hash Hash) (token string, url string, err error) {
-	if h.bridgeDB == nil {
-		err = errors.New("no active bridge")
-		return
-	}
-	err = h.bridgeDB.View(func(tx *buntdb.Tx) (e error) {
-		token, e = tx.Get("app:" + hash.String())
-		if e == buntdb.ErrNotFound {
-			e = BridgeAppNotFoundErr
-		}
-		url, e = tx.Get("url:" + hash.String())
-		if e == buntdb.ErrNotFound {
-			e = BridgeAppNotFoundErr
-		}
-		return
-	})
 	return
 }
